[Security Solution] Remove `unifiedComponentsInTimelineDisabled` flag (#195959)
## Summary

Handles https://github.com/elastic/security-team/issues/9645

Follow Up PR for removal of Old timeline Code :
https://github.com/elastic/kibana/pull/196243

- This PR removes `unifiedComponentsInTimelineDisabled` flag. What this
means that that unified components in Timeline are now enabled by
default.

- Consequently, the old timeline table code becomes obsolete and is also
removed. ( https://github.com/elastic/kibana/pull/196243)

## Changes

1. Converted all cypress and jest tests that were testing old Timeline
table to test new unified components in Timeline. If the test for new
timeline already existed, old tests are also removed.

(cherry picked from commit c41178d2d6e952798548ccd7db691d5ceff62053)
