+++ x-pack/plugins/security_solution/public/common/components/header_actions/header_actions.test.tsx
@@ -14,7 +14,6 @@ import { TimelineTabs } from '../../../../common/types';
 import { HeaderActions } from './header_actions';
 import { timelineActions } from '../../../timelines/store';
 import { getColumnHeader } from '../../../timelines/components/timeline/body/column_headers/helpers';
-import { useIsExperimentalFeatureEnabled } from '../../hooks/use_experimental_features';
 
 jest.mock('../../hooks/use_experimental_features', () => ({
   useIsExperimentalFeatureEnabled: jest.fn(),
@@ -141,51 +140,23 @@ describe('HeaderActions', () => {
     });
   });
 
-  describe('conditional components based on unifiedComponentsInTimelineDisabled', () => {
-    describe('when unifiedComponentsInTimelineDisabled is false', () => {
-      beforeEach(() => {
-        (useIsExperimentalFeatureEnabled as jest.Mock).mockReturnValue(false);
-      });
-      it('should not show the event renderer settings', () => {
-        const result = render(
-          <TestProviders>
-            <HeaderActions {...defaultProps} />
-          </TestProviders>
-        );
-        expect(result.queryByTestId('show-row-renderers-gear')).toBeNull();
-      });
-
-      it('should not show the sorting settings', () => {
-        const result = render(
-          <TestProviders>
-            <HeaderActions {...defaultProps} />
-          </TestProviders>
-        );
-        expect(result.queryByTestId('timeline-sorting-fields')).toBeNull();
-      });
+  describe('Controls', () => {
+    it('should not show the event renderer settings', () => {
+      const result = render(
+        <TestProviders>
+          <HeaderActions {...defaultProps} />
+        </TestProviders>
+      );
+      expect(result.queryByTestId('show-row-renderers-gear')).toBeNull();
     });
 
-    describe('when unifiedComponentsInTimelineDisabled is true', () => {
-      beforeEach(() => {
-        (useIsExperimentalFeatureEnabled as jest.Mock).mockReturnValue(true);
-      });
-      it('should show the event renderer settings', () => {
-        const result = render(
-          <TestProviders>
-            <HeaderActions {...defaultProps} />
-          </TestProviders>
-        );
-        result.getByTestId('show-row-renderers-gear');
-      });
-
-      it('should show the sorting settings', () => {
-        const result = render(
-          <TestProviders>
-            <HeaderActions {...defaultProps} />
-          </TestProviders>
-        );
-        result.getByTestId('timeline-sorting-fields');
-      });
+    it('should not show the sorting settings', () => {
+      const result = render(
+        <TestProviders>
+          <HeaderActions {...defaultProps} />
+        </TestProviders>
+      );
+      expect(result.queryByTestId('timeline-sorting-fields')).toBeNull();
     });
   });
 });
