+++ x-pack/plugins/security_solution/public/common/utils/timeline/use_timeline_click.tsx
@@ -8,14 +8,9 @@
 import { useCallback } from 'react';
 import { useQueryTimelineById } from '../../../timelines/components/open_timeline/helpers';
 import type { TimelineErrorCallback } from '../../../timelines/components/open_timeline/types';
-import { useIsExperimentalFeatureEnabled } from '../../hooks/use_experimental_features';
 
 export const useTimelineClick = () => {
   const queryTimelineById = useQueryTimelineById();
-
-  const unifiedComponentsInTimelineDisabled = useIsExperimentalFeatureEnabled(
-    'unifiedComponentsInTimelineDisabled'
-  );
 
   const handleTimelineClick = useCallback(
     (timelineId: string, onError: TimelineErrorCallback, graphEventId?: string) => {
@@ -23,10 +18,9 @@ export const useTimelineClick = () => {
         graphEventId,
         timelineId,
         onError,
-        unifiedComponentsInTimelineDisabled,
       });
     },
-    [queryTimelineById, unifiedComponentsInTimelineDisabled]
+    [queryTimelineById]
   );
 
   return handleTimelineClick;
