+++ x-pack/plugins/security_solution/public/detections/components/alerts_table/timeline_actions/use_investigate_in_timeline.tsx
@@ -19,7 +19,6 @@ import { useApi } from '@kbn/securitysolution-list-hooks';
 import type { Filter } from '@kbn/es-query';
 import type { EcsSecurityExtension as Ecs } from '@kbn/securitysolution-ecs';
 import { isEmpty } from 'lodash';
-import { useIsExperimentalFeatureEnabled } from '../../../../common/hooks/use_experimental_features';
 import { createHistoryEntry } from '../../../../common/utils/global_query_string/helpers';
 import { useKibana } from '../../../../common/lib/kibana';
 import { TimelineId } from '../../../../../common/types/timeline';
@@ -35,7 +34,6 @@ import { useAppToasts } from '../../../../common/hooks/use_app_toasts';
 import { useStartTransaction } from '../../../../common/lib/apm/use_start_transaction';
 import { ALERTS_ACTIONS } from '../../../../common/lib/apm/user_actions';
 import { defaultUdtHeaders } from '../../../../timelines/components/timeline/unified_components/default_headers';
-import { defaultHeaders } from '../../../../timelines/components/timeline/body/column_headers/default_headers';
 
 interface UseInvestigateInTimelineActionProps {
   ecsRowData?: Ecs | Ecs[] | null;
@@ -146,21 +144,13 @@ export const useInvestigateInTimeline = ({
     timelineType: TimelineTypeEnum.default,
   });
 
-  const unifiedComponentsInTimelineDisabled = useIsExperimentalFeatureEnabled(
-    'unifiedComponentsInTimelineDisabled'
-  );
-
   const updateTimeline = useUpdateTimeline();
 
   const createTimeline = useCallback(
     async ({ from: fromTimeline, timeline, to: toTimeline, ruleNote }: CreateTimelineProps) => {
       const newColumns = timeline.columns;
       const newColumnsOverride =
-        !newColumns || isEmpty(newColumns)
-          ? !unifiedComponentsInTimelineDisabled
-            ? defaultUdtHeaders
-            : defaultHeaders
-          : newColumns;
+        !newColumns || isEmpty(newColumns) ? defaultUdtHeaders : newColumns;
 
       await clearActiveTimeline();
       updateTimelineIsLoading({ id: TimelineId.active, isLoading: false });
@@ -175,7 +165,6 @@ export const useInvestigateInTimeline = ({
           indexNames: timeline.indexNames ?? [],
           show: true,
           excludedRowRendererIds:
-            !unifiedComponentsInTimelineDisabled &&
             timeline.timelineType !== TimelineTypeEnum.template
               ? timeline.excludedRowRendererIds
               : [],
@@ -184,12 +173,7 @@ export const useInvestigateInTimeline = ({
         ruleNote,
       });
     },
-    [
-      updateTimeline,
-      updateTimelineIsLoading,
-      clearActiveTimeline,
-      unifiedComponentsInTimelineDisabled,
-    ]
+    [updateTimeline, updateTimelineIsLoading, clearActiveTimeline]
   );
 
   const investigateInTimelineAlertClick = useCallback(async () => {
