+++ x-pack/plugins/security_solution/public/notes/components/open_timeline_button.tsx
@@ -8,7 +8,6 @@
 import React, { memo, useCallback } from 'react';
 import { EuiButtonIcon } from '@elastic/eui';
 import { i18n } from '@kbn/i18n';
-import { useIsExperimentalFeatureEnabled } from '../../common/hooks/use_experimental_features';
 import { useQueryTimelineById } from '../../timelines/components/open_timeline/helpers';
 import { OPEN_TIMELINE_BUTTON_TEST_ID } from './test_ids';
 import type { Note } from '../../../common/api/timeline';
@@ -32,10 +31,6 @@ export interface OpenTimelineButtonIconProps {
  * Renders a button to open the timeline associated with a note
  */
 export const OpenTimelineButtonIcon = memo(({ note, index }: OpenTimelineButtonIconProps) => {
-  const unifiedComponentsInTimelineDisabled = useIsExperimentalFeatureEnabled(
-    'unifiedComponentsInTimelineDisabled'
-  );
-
   const queryTimelineById = useQueryTimelineById();
   const openTimeline = useCallback(
     ({ timelineId }: { timelineId: string }) =>
@@ -44,9 +39,8 @@ export const OpenTimelineButtonIcon = memo(({ note, index }: OpenTimelineButtonI
         onOpenTimeline: undefined,
         timelineId,
         timelineType: undefined,
-        unifiedComponentsInTimelineDisabled,
       }),
-    [queryTimelineById, unifiedComponentsInTimelineDisabled]
+    [queryTimelineById]
   );
 
   return (
