+++ x-pack/plugins/security_solution/public/timelines/components/modal/actions/new_timeline_button.test.tsx
@@ -10,9 +10,7 @@ import React from 'react';
 import { NewTimelineButton } from './new_timeline_button';
 import { TimelineId } from '../../../../../common/types';
 import { timelineActions } from '../../../store';
-import { defaultHeaders } from '../../timeline/body/column_headers/default_headers';
 import { TestProviders } from '../../../../common/mock';
-import { useIsExperimentalFeatureEnabled } from '../../../../common/hooks/use_experimental_features';
 import { RowRendererValues } from '../../../../../common/api/timeline';
 import { defaultUdtHeaders } from '../../timeline/unified_components/default_headers';
 
@@ -74,27 +72,6 @@ describe('NewTimelineButton', () => {
         timelineType: 'default',
         updated: undefined,
         excludedRowRendererIds: RowRendererValues,
-      });
-    });
-
-    // disable unified components in timeline
-    (useIsExperimentalFeatureEnabled as jest.Mock).mockReturnValue(true);
-
-    getByTestId('timeline-modal-new-timeline-dropdown-button').click();
-    getByTestId('timeline-modal-new-timeline').click();
-
-    spy.mockClear();
-
-    await waitFor(() => {
-      expect(spy).toHaveBeenCalledWith({
-        columns: defaultHeaders,
-        dataViewId,
-        id: TimelineId.test,
-        indexNames: selectedPatterns,
-        show: true,
-        timelineType: 'default',
-        updated: undefined,
-        excludedRowRendererIds: [],
       });
     });
   });
