+++ x-pack/plugins/security_solution/public/timelines/components/timeline/tabs/eql/index.tsx
@@ -17,23 +17,17 @@ import type { EuiDataGridControlColumn } from '@elastic/eui';
 import { DataLoadingState } from '@kbn/unified-data-table';
 import { useExpandableFlyoutApi } from '@kbn/expandable-flyout';
 import type { RunTimeMappings } from '@kbn/timelines-plugin/common/search_strategy';
+import { InputsModelId } from '../../../../../common/store/inputs/constants';
 import { useKibana } from '../../../../../common/lib/kibana';
 import {
   DocumentDetailsLeftPanelKey,
   DocumentDetailsRightPanelKey,
 } from '../../../../../flyout/document_details/shared/constants/panel_keys';
-import { InputsModelId } from '../../../../../common/store/inputs/constants';
-import type { ControlColumnProps } from '../../../../../../common/types';
 import { useDeepEqualSelector } from '../../../../../common/hooks/use_selector';
 import { useIsExperimentalFeatureEnabled } from '../../../../../common/hooks/use_experimental_features';
 import { timelineActions, timelineSelectors } from '../../../../store';
 import { useTimelineEvents } from '../../../../containers';
-import { StatefulBody } from '../../body';
-import { Footer, footerHeight } from '../../footer';
-import { calculateTotalPages } from '../../helpers';
-import { TimelineRefetch } from '../../refetch_timeline';
 import { TimelineId, TimelineTabs } from '../../../../../../common/types/timeline';
-import { EventDetailsWidthProvider } from '../../../../../common/components/events_viewer/event_details_width_context';
 import type { inputsModel, State } from '../../../../../common/store';
 import { inputsSelectors } from '../../../../../common/store';
 import { SourcererScopeName } from '../../../../../sourcerer/store/model';
@@ -42,19 +36,8 @@ import { useSourcererDataView } from '../../../../../sourcerer/containers';
 import { useEqlEventsCountPortal } from '../../../../../common/hooks/use_timeline_events_count';
 import type { TimelineModel } from '../../../../store/model';
 import { useTimelineFullScreen } from '../../../../../common/containers/use_full_screen';
-import {
-  EventsCountBadge,
-  FullWidthFlexGroup,
-  ScrollableFlexItem,
-  StyledEuiFlyoutBody,
-  StyledEuiFlyoutFooter,
-} from '../shared/layout';
-import {
-  TIMELINE_EMPTY_EVENTS,
-  isTimerangeSame,
-  timelineEmptyTrailingControlColumns,
-  TIMELINE_NO_SORTING,
-} from '../shared/utils';
+import { EventsCountBadge, FullWidthFlexGroup } from '../shared/layout';
+import { isTimerangeSame, TIMELINE_NO_SORTING } from '../shared/utils';
 import type { TimelineTabCommonProps } from '../shared/types';
 import { UnifiedTimelineBody } from '../../body/unified_timeline_body';
 import { EqlTabHeader } from './header';
@@ -63,6 +46,7 @@ import { useTimelineControlColumn } from '../shared/use_timeline_control_columns
 import { LeftPanelNotesTab } from '../../../../../flyout/document_details/left';
 import { useNotesInFlyout } from '../../properties/use_notes_in_flyout';
 import { NotesFlyout } from '../../properties/notes_flyout';
+import { TimelineRefetch } from '../../refetch_timeline';
 
 export type Props = TimelineTabCommonProps & PropsFromRedux;
 
@@ -72,10 +56,8 @@ export const EqlTabContentComponent: React.FC<Props> = ({
   end,
   eqlOptions,
   timelineId,
-  isLive,
   itemsPerPage,
   itemsPerPageOptions,
-  renderCellValue,
   rowRenderers,
   start,
   timerangeKind,
@@ -88,17 +70,12 @@ export const EqlTabContentComponent: React.FC<Props> = ({
   const { portalNode: eqlEventsCountPortalNode } = useEqlEventsCountPortal();
   const { setTimelineFullScreen, timelineFullScreen } = useTimelineFullScreen();
   const {
-    browserFields,
     dataViewId,
     loading: loadingSourcerer,
     selectedPatterns,
     sourcererDataView,
   } = useSourcererDataView(SourcererScopeName.timeline);
   const { augmentedColumnHeaders, timelineQueryFieldsFromColumns } = useTimelineColumns(columns);
-
-  const unifiedComponentsInTimelineDisabled = useIsExperimentalFeatureEnabled(
-    'unifiedComponentsInTimelineDisabled'
-  );
 
   const getManageTimeline = useMemo(() => timelineSelectors.getTimelineByIdSelector(), []);
 
@@ -132,7 +109,7 @@ export const EqlTabContentComponent: React.FC<Props> = ({
     id: timelineId,
     indexNames: selectedPatterns,
     language: 'eql',
-    limit: !unifiedComponentsInTimelineDisabled ? sampleSize : itemsPerPage,
+    limit: sampleSize,
     runtimeMappings: sourcererDataView?.runtimeFieldMap as RunTimeMappings,
     skip: !canQueryTimeline(),
     startDate: start,
@@ -267,103 +244,42 @@ export const EqlTabContentComponent: React.FC<Props> = ({
 
   return (
     <>
-      {!unifiedComponentsInTimelineDisabled ? (
-        <>
-          <InPortal node={eqlEventsCountPortalNode}>
-            {totalCount >= 0 ? (
-              <EventsCountBadge data-test-subj="eql-events-count">{totalCount}</EventsCountBadge>
-            ) : null}
-          </InPortal>
-          {NotesFlyoutMemo}
-          <FullWidthFlexGroup>
-            <UnifiedTimelineBody
-              header={unifiedHeader}
-              columns={augmentedColumnHeaders}
-              isSortEnabled={false}
-              rowRenderers={rowRenderers}
-              timelineId={timelineId}
-              itemsPerPage={itemsPerPage}
-              itemsPerPageOptions={itemsPerPageOptions}
-              sort={TIMELINE_NO_SORTING}
-              events={events}
-              refetch={refetch}
-              dataLoadingState={dataLoadingState}
-              totalCount={isBlankTimeline ? 0 : totalCount}
-              onChangePage={loadPage}
-              activeTab={activeTab}
-              updatedAt={refreshedAt}
-              isTextBasedQuery={false}
-              pageInfo={pageInfo}
-              leadingControlColumns={leadingControlColumns as EuiDataGridControlColumn[]}
-            />
-          </FullWidthFlexGroup>
-        </>
-      ) : (
-        <>
-          <InPortal node={eqlEventsCountPortalNode}>
-            {totalCount >= 0 ? <EventsCountBadge>{totalCount}</EventsCountBadge> : null}
-          </InPortal>
-          {NotesFlyoutMemo}
-          <TimelineRefetch
-            id={`${timelineId}-${TimelineTabs.eql}`}
-            inputId={InputsModelId.timeline}
-            inspect={inspect}
-            loading={isQueryLoading}
-            refetch={refetch}
-          />
-          <FullWidthFlexGroup gutterSize="s" direction="column">
-            <ScrollableFlexItem grow={false}>{unifiedHeader}</ScrollableFlexItem>
-            <ScrollableFlexItem grow={true}>
-              <EventDetailsWidthProvider>
-                <StyledEuiFlyoutBody
-                  data-test-subj={`${TimelineTabs.eql}-tab-flyout-body`}
-                  className="timeline-flyout-body"
-                >
-                  <StatefulBody
-                    activePage={pageInfo.activePage}
-                    browserFields={browserFields}
-                    data={isBlankTimeline ? TIMELINE_EMPTY_EVENTS : events}
-                    id={timelineId}
-                    refetch={refetch}
-                    renderCellValue={renderCellValue}
-                    rowRenderers={rowRenderers}
-                    sort={TIMELINE_NO_SORTING}
-                    tabType={TimelineTabs.eql}
-                    totalPages={calculateTotalPages({
-                      itemsCount: totalCount,
-                      itemsPerPage,
-                    })}
-                    leadingControlColumns={leadingControlColumns as ControlColumnProps[]}
-                    trailingControlColumns={timelineEmptyTrailingControlColumns}
-                  />
-                </StyledEuiFlyoutBody>
-
-                <StyledEuiFlyoutFooter
-                  data-test-subj={`${TimelineTabs.eql}-tab-flyout-footer`}
-                  className="timeline-flyout-footer"
-                >
-                  {!isBlankTimeline && (
-                    <Footer
-                      activePage={pageInfo?.activePage ?? 0}
-                      data-test-subj="timeline-footer"
-                      updatedAt={refreshedAt}
-                      height={footerHeight}
-                      id={timelineId}
-                      isLive={isLive}
-                      isLoading={isQueryLoading || loadingSourcerer}
-                      itemsCount={isBlankTimeline ? 0 : events.length}
-                      itemsPerPage={itemsPerPage}
-                      itemsPerPageOptions={itemsPerPageOptions}
-                      onChangePage={loadPage}
-                      totalCount={isBlankTimeline ? 0 : totalCount}
-                    />
-                  )}
-                </StyledEuiFlyoutFooter>
-              </EventDetailsWidthProvider>
-            </ScrollableFlexItem>
-          </FullWidthFlexGroup>
-        </>
-      )}
+      <TimelineRefetch
+        id={`${timelineId}-${TimelineTabs.eql}`}
+        inputId={InputsModelId.timeline}
+        inspect={inspect}
+        loading={isQueryLoading}
+        refetch={refetch}
+        skip={!canQueryTimeline}
+      />
+      <InPortal node={eqlEventsCountPortalNode}>
+        {totalCount >= 0 ? (
+          <EventsCountBadge data-test-subj="eql-events-count">{totalCount}</EventsCountBadge>
+        ) : null}
+      </InPortal>
+      {NotesFlyoutMemo}
+      <FullWidthFlexGroup>
+        <UnifiedTimelineBody
+          header={unifiedHeader}
+          columns={augmentedColumnHeaders}
+          isSortEnabled={false}
+          rowRenderers={rowRenderers}
+          timelineId={timelineId}
+          itemsPerPage={itemsPerPage}
+          itemsPerPageOptions={itemsPerPageOptions}
+          sort={TIMELINE_NO_SORTING}
+          events={events}
+          refetch={refetch}
+          dataLoadingState={dataLoadingState}
+          totalCount={isBlankTimeline ? 0 : totalCount}
+          onChangePage={loadPage}
+          activeTab={activeTab}
+          updatedAt={refreshedAt}
+          isTextBasedQuery={false}
+          pageInfo={pageInfo}
+          leadingControlColumns={leadingControlColumns as EuiDataGridControlColumn[]}
+        />
+      </FullWidthFlexGroup>
     </>
   );
 };
