+++ x-pack/plugins/security_solution/public/timelines/components/timeline/tabs/query/index.tsx
@@ -27,18 +27,13 @@ import { InputsModelId } from '../../../../../common/store/inputs/constants';
 import { useInvalidFilterQuery } from '../../../../../common/hooks/use_invalid_filter_query';
 import { timelineActions, timelineSelectors } from '../../../../store';
 import type { Direction } from '../../../../../../common/search_strategy';
-import type { ControlColumnProps } from '../../../../../../common/types';
 import { useTimelineEvents } from '../../../../containers';
 import { useKibana } from '../../../../../common/lib/kibana';
-import { StatefulBody } from '../../body';
-import { Footer, footerHeight } from '../../footer';
 import { QueryTabHeader } from './header';
-import { calculateTotalPages } from '../../helpers';
 import { combineQueries } from '../../../../../common/lib/kuery';
 import { TimelineRefetch } from '../../refetch_timeline';
 import type { KueryFilterQueryKind } from '../../../../../../common/types/timeline';
 import { TimelineId, TimelineTabs } from '../../../../../../common/types/timeline';
-import { EventDetailsWidthProvider } from '../../../../../common/components/events_viewer/event_details_width_context';
 import type { inputsModel, State } from '../../../../../common/store';
 import { inputsSelectors } from '../../../../../common/store';
 import { SourcererScopeName } from '../../../../../sourcerer/store/model';
@@ -48,17 +43,7 @@ import { isActiveTimeline } from '../../../../../helpers';
 
 import type { TimelineModel } from '../../../../store/model';
 import { UnifiedTimelineBody } from '../../body/unified_timeline_body';
-import {
-  FullWidthFlexGroup,
-  ScrollableFlexItem,
-  StyledEuiFlyoutBody,
-  StyledEuiFlyoutFooter,
-} from '../shared/layout';
-import {
-  TIMELINE_EMPTY_EVENTS,
-  isTimerangeSame,
-  timelineEmptyTrailingControlColumns,
-} from '../shared/utils';
+import { isTimerangeSame } from '../shared/utils';
 import type { TimelineTabCommonProps } from '../shared/types';
 import { useTimelineColumns } from '../shared/use_timeline_columns';
 import { useTimelineControlColumn } from '../shared/use_timeline_control_columns';
@@ -112,10 +97,6 @@ export const QueryTabContentComponent: React.FC<Props> = ({
   const {
     query: { filterManager: timelineFilterManager },
   } = timelineDataService;
-
-  const unifiedComponentsInTimelineDisabled = useIsExperimentalFeatureEnabled(
-    'unifiedComponentsInTimelineDisabled'
-  );
 
   const getManageTimeline = useMemo(() => timelineSelectors.getTimelineByIdSelector(), []);
 
@@ -195,7 +176,7 @@ export const QueryTabContentComponent: React.FC<Props> = ({
     id: timelineId,
     indexNames: selectedPatterns,
     language: kqlQuery.language,
-    limit: !unifiedComponentsInTimelineDisabled ? sampleSize : itemsPerPage,
+    limit: sampleSize,
     runtimeMappings: sourcererDataView?.runtimeFieldMap as RunTimeMappings,
     skip: !canQueryTimeline,
     sort: timelineQuerySortField,
@@ -347,53 +328,6 @@ export const QueryTabContentComponent: React.FC<Props> = ({
     );
   }, [associateNote, closeNotesFlyout, isNotesFlyoutVisible, noteEventId, notes, timelineId]);
 
-  if (!unifiedComponentsInTimelineDisabled) {
-    return (
-      <>
-        <TimelineRefetch
-          id={`${timelineId}-${TimelineTabs.query}`}
-          inputId={InputsModelId.timeline}
-          inspect={inspect}
-          loading={isQueryLoading}
-          refetch={refetch}
-          skip={!canQueryTimeline}
-        />
-        {NotesFlyoutMemo}
-
-        <UnifiedTimelineBody
-          header={
-            <QueryTabHeader
-              activeTab={activeTab}
-              filterManager={timelineFilterManager}
-              show={show && activeTab === TimelineTabs.query}
-              showCallOutUnauthorizedMsg={showCallOutUnauthorizedMsg}
-              status={status}
-              timelineId={timelineId}
-              showEventsCountBadge={showEventsCountBadge}
-              totalCount={totalCount}
-            />
-          }
-          columns={augmentedColumnHeaders}
-          rowRenderers={rowRenderers}
-          timelineId={timelineId}
-          itemsPerPage={itemsPerPage}
-          itemsPerPageOptions={itemsPerPageOptions}
-          sort={sort}
-          events={events}
-          refetch={refetch}
-          dataLoadingState={dataLoadingState}
-          totalCount={isBlankTimeline ? 0 : totalCount}
-          leadingControlColumns={leadingControlColumns as EuiDataGridControlColumn[]}
-          onChangePage={loadPage}
-          activeTab={activeTab}
-          updatedAt={refreshedAt}
-          isTextBasedQuery={false}
-          pageInfo={pageInfo}
-        />
-      </>
-    );
-  }
-
   return (
     <>
       <TimelineRefetch
@@ -405,8 +339,9 @@ export const QueryTabContentComponent: React.FC<Props> = ({
         skip={!canQueryTimeline}
       />
       {NotesFlyoutMemo}
-      <FullWidthFlexGroup gutterSize="none">
-        <ScrollableFlexItem grow={2}>
+
+      <UnifiedTimelineBody
+        header={
           <QueryTabHeader
             activeTab={activeTab}
             filterManager={timelineFilterManager}
@@ -417,55 +352,24 @@ export const QueryTabContentComponent: React.FC<Props> = ({
             showEventsCountBadge={showEventsCountBadge}
             totalCount={totalCount}
           />
-          <EventDetailsWidthProvider>
-            <StyledEuiFlyoutBody
-              data-test-subj={`${TimelineTabs.query}-tab-flyout-body`}
-              className="timeline-flyout-body"
-            >
-              <StatefulBody
-                activePage={pageInfo.activePage}
-                browserFields={browserFields}
-                data={isBlankTimeline ? TIMELINE_EMPTY_EVENTS : events}
-                id={timelineId}
-                refetch={refetch}
-                renderCellValue={renderCellValue}
-                rowRenderers={rowRenderers}
-                sort={sort}
-                tabType={TimelineTabs.query}
-                totalPages={calculateTotalPages({
-                  itemsCount: totalCount,
-                  itemsPerPage,
-                })}
-                leadingControlColumns={leadingControlColumns as ControlColumnProps[]}
-                trailingControlColumns={timelineEmptyTrailingControlColumns}
-                onToggleShowNotes={onToggleShowNotes}
-              />
-            </StyledEuiFlyoutBody>
-
-            <StyledEuiFlyoutFooter
-              data-test-subj={`${TimelineTabs.query}-tab-flyout-footer`}
-              className="timeline-flyout-footer"
-            >
-              {!isBlankTimeline && (
-                <Footer
-                  activePage={pageInfo?.activePage ?? 0}
-                  data-test-subj="timeline-footer"
-                  updatedAt={refreshedAt}
-                  height={footerHeight}
-                  id={timelineId}
-                  isLive={isLive}
-                  isLoading={isQueryLoading || loadingSourcerer}
-                  itemsCount={isBlankTimeline ? 0 : events.length}
-                  itemsPerPage={itemsPerPage}
-                  itemsPerPageOptions={itemsPerPageOptions}
-                  onChangePage={loadPage}
-                  totalCount={isBlankTimeline ? 0 : totalCount}
-                />
-              )}
-            </StyledEuiFlyoutFooter>
-          </EventDetailsWidthProvider>
-        </ScrollableFlexItem>
-      </FullWidthFlexGroup>
+        }
+        columns={augmentedColumnHeaders}
+        rowRenderers={rowRenderers}
+        timelineId={timelineId}
+        itemsPerPage={itemsPerPage}
+        itemsPerPageOptions={itemsPerPageOptions}
+        sort={sort}
+        events={events}
+        refetch={refetch}
+        dataLoadingState={dataLoadingState}
+        totalCount={isBlankTimeline ? 0 : totalCount}
+        leadingControlColumns={leadingControlColumns as EuiDataGridControlColumn[]}
+        onChangePage={loadPage}
+        activeTab={activeTab}
+        updatedAt={refreshedAt}
+        isTextBasedQuery={false}
+        pageInfo={pageInfo}
+      />
     </>
   );
 };
