+++ x-pack/plugins/security_solution/public/timelines/components/timeline/tabs/shared/use_timeline_columns.tsx
@@ -8,8 +8,6 @@
 import { useMemo } from 'react';
 import { SourcererScopeName } from '../../../../../sourcerer/store/model';
 import { useSourcererDataView } from '../../../../../sourcerer/containers';
-import { useIsExperimentalFeatureEnabled } from '../../../../../common/hooks/use_experimental_features';
-import { defaultHeaders } from '../../body/column_headers/default_headers';
 import { requiredFieldsForActions } from '../../../../../detections/components/alerts_table/default_config';
 import { defaultUdtHeaders } from '../../unified_components/default_headers';
 import type { ColumnHeaderOptions } from '../../../../../../common/types';
@@ -18,16 +16,7 @@ import { memoizedGetTimelineColumnHeaders } from './utils';
 export const useTimelineColumns = (columns: ColumnHeaderOptions[]) => {
   const { browserFields } = useSourcererDataView(SourcererScopeName.timeline);
 
-  const unifiedComponentsInTimelineDisabled = useIsExperimentalFeatureEnabled(
-    'unifiedComponentsInTimelineDisabled'
-  );
-
-  const defaultColumns = useMemo(
-    () => (!unifiedComponentsInTimelineDisabled ? defaultUdtHeaders : defaultHeaders),
-    [unifiedComponentsInTimelineDisabled]
-  );
-
-  const localColumns = useMemo(() => columns ?? defaultColumns, [columns, defaultColumns]);
+  const localColumns = useMemo(() => columns ?? defaultUdtHeaders, [columns]);
 
   const augmentedColumnHeaders = memoizedGetTimelineColumnHeaders(
     localColumns,
@@ -43,11 +32,11 @@ export const useTimelineColumns = (columns: ColumnHeaderOptions[]) => {
 
   return useMemo(
     () => ({
-      defaultColumns,
+      defaultColumns: defaultUdtHeaders,
       localColumns,
       augmentedColumnHeaders,
       timelineQueryFieldsFromColumns,
     }),
-    [augmentedColumnHeaders, defaultColumns, timelineQueryFieldsFromColumns, localColumns]
+    [augmentedColumnHeaders, timelineQueryFieldsFromColumns, localColumns]
   );
 };
