+++ x-pack/plugins/security_solution/public/timelines/components/timeline/tabs/shared/use_timeline_control_columns.test.tsx
@@ -21,7 +21,7 @@ jest.mock('../../../../../common/hooks/use_license', () => ({
 
 const useLicenseMock = useLicense as jest.Mock;
 
-describe('useTimelineColumns', () => {
+describe('useTimelineControlColumns', () => {
   const mockColumns: ColumnHeaderOptions[] = [
     {
       columnHeaderType: 'not-filtered',
