+++ x-pack/plugins/security_solution/public/timelines/components/timeline/tabs/shared/use_timeline_control_columns.tsx
@@ -13,7 +13,6 @@ import { JEST_ENVIRONMENT } from '../../../../../../common/constants';
 import { useLicense } from '../../../../../common/hooks/use_license';
 import { SourcererScopeName } from '../../../../../sourcerer/store/model';
 import { useSourcererDataView } from '../../../../../sourcerer/containers';
-import { useIsExperimentalFeatureEnabled } from '../../../../../common/hooks/use_experimental_features';
 import { getDefaultControlColumn } from '../../body/control_columns';
 import type { UnifiedActionProps } from '../../unified_components/data_table/control_column_cell_render';
 import type { TimelineTabs } from '../../../../../../common/types/timeline';
@@ -53,12 +52,8 @@ export const useTimelineControlColumn = ({
 }: UseTimelineControlColumnArgs) => {
   const { browserFields } = useSourcererDataView(SourcererScopeName.timeline);
 
-  const unifiedComponentsInTimelineDisabled = useIsExperimentalFeatureEnabled(
-    'unifiedComponentsInTimelineDisabled'
-  );
-
   const isEnterprisePlus = useLicense().isEnterprise();
-  const ACTION_BUTTON_COUNT = isEnterprisePlus ? 6 : 5;
+  const ACTION_BUTTON_COUNT = useMemo(() => (isEnterprisePlus ? 6 : 5), [isEnterprisePlus]);
   const { localColumns } = useTimelineColumns(columns);
 
   const RowCellRender = useMemo(
@@ -116,45 +111,36 @@ export const useTimelineControlColumn = ({
   // We need one less when the unified components are enabled because the document expand is provided by the unified data table
   const UNIFIED_COMPONENTS_ACTION_BUTTON_COUNT = ACTION_BUTTON_COUNT - 1;
   return useMemo(() => {
-    if (!unifiedComponentsInTimelineDisabled) {
-      return getDefaultControlColumn(UNIFIED_COMPONENTS_ACTION_BUTTON_COUNT).map((x) => ({
-        ...x,
-        headerCellRender: function HeaderCellRender(props: UnifiedActionProps) {
-          return (
-            <HeaderActions
-              width={x.width}
-              browserFields={browserFields}
-              columnHeaders={localColumns}
-              isEventViewer={false}
-              isSelectAllChecked={false}
-              onSelectAll={noSelectAll}
-              showEventsSelect={false}
-              showSelectAllCheckbox={false}
-              showFullScreenToggle={false}
-              sort={sort}
-              tabType={activeTab}
-              {...props}
-              timelineId={timelineId}
-            />
-          );
-        },
-        rowCellRender: JEST_ENVIRONMENT ? RowCellRender : React.memo(RowCellRender),
-      }));
-    } else {
-      return getDefaultControlColumn(ACTION_BUTTON_COUNT).map((x) => ({
-        ...x,
-        headerCellRender: HeaderActions,
-      })) as unknown as ColumnHeaderOptions[];
-    }
+    return getDefaultControlColumn(UNIFIED_COMPONENTS_ACTION_BUTTON_COUNT).map((x) => ({
+      ...x,
+      headerCellRender: function HeaderCellRender(props: UnifiedActionProps) {
+        return (
+          <HeaderActions
+            width={x.width}
+            browserFields={browserFields}
+            columnHeaders={localColumns}
+            isEventViewer={false}
+            isSelectAllChecked={false}
+            onSelectAll={noSelectAll}
+            showEventsSelect={false}
+            showSelectAllCheckbox={false}
+            showFullScreenToggle={false}
+            sort={sort}
+            tabType={activeTab}
+            {...props}
+            timelineId={timelineId}
+          />
+        );
+      },
+      rowCellRender: JEST_ENVIRONMENT ? RowCellRender : React.memo(RowCellRender),
+    }));
   }, [
-    unifiedComponentsInTimelineDisabled,
     UNIFIED_COMPONENTS_ACTION_BUTTON_COUNT,
     browserFields,
     localColumns,
     sort,
     activeTab,
     timelineId,
-    ACTION_BUTTON_COUNT,
     RowCellRender,
   ]);
 };
