+++ x-pack/plugins/security_solution/public/timelines/components/timeline/unified_components/index.test.tsx
@@ -67,9 +67,6 @@ jest.mock('react-router-dom', () => ({
 }));
 
 const useIsExperimentalFeatureEnabledMock = jest.fn((feature: keyof ExperimentalFeatures) => {
-  if (feature === 'unifiedComponentsInTimelineDisabled') {
-    return false;
-  }
   return allowedExperimentalValues[feature];
 });
